Changed the project to Android Studio
Android Studio is a nicer development environment, and is really easy to
set up and get phones working with.  This new branch of the project
contains all the files to have the whole project as it is designed for
Studio.  I'd prefer working here, but if people really don't want to I
can just manually add the updated files.